@@ -1,17 +1,13 @@
 ﻿using Books.Models;
-using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
 
-namespace BooksApi.Services
+namespace BooksApi.Services;
+
+public interface IBookChapterService
 {
-    public interface IBookChapterService
-    {
-        Task AddAsync(BookChapter chapter);
-        Task AddRangeAsync(IEnumerable<BookChapter> chapters);
-        Task<IEnumerable<BookChapter>> GetAllAsync();
-        Task<BookChapter?> FindAsync(Guid id);
-        Task<BookChapter?> RemoveAsync(Guid id);
-        Task<BookChapter?> UpdateAsync(BookChapter chapter);
-    }
+    Task AddAsync(BookChapter chapter);
+    Task AddRangeAsync(IEnumerable<BookChapter> chapters);
+    Task<IEnumerable<BookChapter>> GetAllAsync();
+    Task<BookChapter?> FindAsync(Guid id);
+    Task<BookChapter?> RemoveAsync(Guid id);
+    Task<BookChapter?> UpdateAsync(BookChapter chapter);
 }
