@@ -1,20 +1,39 @@
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Hosting;
+using BooksApi.Services;
+using Microsoft.OpenApi.Models;
 
-namespace BooksApi
+var builder = WebApplication.CreateBuilder(args);
+
+builder.Services.AddControllers();
+builder.Services.AddEndpointsApiExplorer();
+
+// builder.Services.AddSwaggerGen();
+builder.Services.AddSwaggerGen(c =>
 {
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            CreateHostBuilder(args).Build().Run();
-        }
+    c.SwaggerDoc("v4", new OpenApiInfo { Title = "BooksApi", Version = "v3" });
+});
 
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureWebHostDefaults(webBuilder =>
-                {
-                    webBuilder.UseStartup<Startup>();
-                });
-    }
+builder.Services.AddSingleton<IBookChapterService, BookChapterService>();
+builder.Services.AddScoped<SampleChapters>();
+
+var app = builder.Build();
+
+if (app.Environment.IsDevelopment())
+{
+    app.UseSwagger();
+    app.UseSwaggerUI(options =>
+    {
+        options.SwaggerEndpoint("/swagger/v4/swagger.json", "Books API");
+    });
 }
+
+app.UseHttpsRedirection();
+app.UseAuthentication();
+app.MapControllers();
+
+app.MapGet("/init", async (SampleChapters sampleChapters, HttpContext context) =>
+{
+    sampleChapters.CreateSampleChapters();
+    await context.Response.WriteAsync("sample chapters initialized");
+});
+app.Run();
+
